Ensure Tools->External plugins is populated
Since the commit b15ed221bd6653c3f01f0ec4754ef42f2e6b4b80, the menu tools->external plugins is empty.
This is because on new system, an added submenu is cloned when added. So any update  on the object
after Add it as no effect. The menu stay empty.
With this commit the menu filling is made before adding submenu.