Use a bigger hammer with wxWidgets.
As if the original hack wasn't embarassing enough, the only change
here is to make the number bigger.

WICKED HILLARY wanted to use wxBoxSizers that are FAR TOO SMALL.
So sad.
But we're going to look at that.  I think we're going to have
BIG wxBoxSizers.  Probably the BIGGEST wxBoxSizers of anyone.

(The bug is that the Sheet Border swatch in PANEL_EESCHEMA_SETTINGS
doesn't otherwise get drawn unless that panel is the one showing
when preferences is opened.)