Kicad: fix environment variable changes after running Pcbnew in some cases.
When pcbnew is built with python support, the env vars defined inside Kicad
(and used mainly in paths) are rewritten for the python environment.
Unfortunately, the values are rewritten as UTF8 string, but in Kicad they are
just expected in utf32 unicode values.

This is now fixed by rewriting the initial values in Kicad env, after the
python environment is initialized.