Fix potential bug in 3D model selection dialog.
Calling EndModal or EndQuasiModal will cause KiCad to hang if the dialog
was not opened in the same manor.  The modality of the dialog must be
checked and the appropriate end method must be called or emulate a button
click which allows the default dialog close method to be called to handle
shutting down the dialog correctly.