Eeschema: Adds wire when dragging directly connected labels
CHANGED: KiCAD allows connecting components directly together in
schematics.  If the user drags one of the components, KiCAD
adds a wire to keep the conection intact.  It should do the same for
net and global labels but it doesn't.

This MR adds net and global labels to the list of items that can connect
to other items.

Fixes https://gitlab.com/kicad/code/kicad/issues/7713