Specify the icon for the file types
For some reason, the icon wasn't always being associated with the files,
so we can force an icon for the files instead. Note that we can't use
the <icon> tag for this, because that is user-space only, so we have
to make the icons be the generic icons to allow for system-level usage.