Fix printing of hidden refs and/or values.
IsLayerVisible() is patched to always return true when printing
(even when "print according to appearance settings" is checked, and
that's too risky to change at this point.  So we also look to see
if the layer colour is set to the background colour.

Fixes https://gitlab.com/kicad/code/kicad/issues/8548