Setup secondary LINGUAS to allow partial install
Prior to release, we will disable languages that don't meet our criteria
(minimum translation > 75%) but we want to keep the languages available
for translation.  LINGUAS lists all languages available.
LINGUAS_INSTALL lists just the languages that will be installed for use
in KiCad.  Right now they are the same content.  This will be changed
prior to v6 release